ReviewState -> ENUM
The ReviewState had been changed to an Enum (this might have caused
some problems I cannot yet foresee). Also the Crtl+ commands in the
menu items have been changed to Meta+ (I hope this is Crtl for Windows
(and I know it is cmd for Mac)).
Also some bugFixes and the first setup for the MessageDialog has been
implemented.
The Open and New menuItems in the File menu have been implemented (the
File menu has been renamed to Project menu).
And some more bug fixes.